requirements.txt
allure-pytest
pytest
